### 2017-09-18 #406 @whedon
[PRE REVIEW]: MatDL: A Lightweight Deep Learning Library in MATLAB
**Submitting author:** @haythamfayek (<a href="http://orcid.org/0000-0002-1840-7605">Haytham Fayek</a>)
**Repository:** <a href="https://github.com/haythamfayek/MatDL" target ="_blank">https://github.com/haythamfayek/MatDL</a>
**Version:** v0.7.0
