### 2017-09-20 #406 @haythamfayek
Thanks @cMadan! From the list of [reviewers](https://github.com/openjournals/joss/blob/master/docs/reviewers.csv), how about @Kevin-Mattheus-Moerman, @nirum, or @sealhuang?
