### 2017-09-21 #406 @sealhuang
@haythamfayek Thank you for mentioning me. I am not very familiar with the deep learning, so if that's not a problem I can do the review. And if someone with more DL experience would like to handle this review, I think it may be more helpful for this software.
