### 2017-09-22 #406 @nirum
This looks interesting! I'm happy to review it
