### 2017-11-01 haythamfayek/MatDL#1 @nirum
MNIST data
One slightly annoying thing about the CNN demo is that I don't know what format the MNIST data is in (`mnist_uint8.mat`). It would be nice if you could add either the data file itself (hosted somewhere) or a script that downloads MNIST data from the web and converts it into this mat file format.

(would just make it easier to quickly run your demo)
