### 2017-11-01 haythamfayek/MatDL#2 @nirum
Add community guidelines
Per JOSS requirements, please add "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" to the README
