### 2017-11-01 #413 @nirum
p.s. here's my CNN training on MNIST 😃
