### 2017-11-01 #413 @nirum
One other comment--there aren't tests that can be run--but there is gradient checking code, and a number of pre-specified architectures. I would encourage @haythamfayek to include some tests for his code (perhaps there are some ideas to be harvested from [this post](https://medium.com/@keeper6928/how-to-unit-test-machine-learning-code-57cf6fd81765)), but I don't think that this is a blocking issue.
