### 2017-11-01 haythamfayek/MatDL#2 @haythamfayek
Added Contributing Guidelines to cover the 3 requirements to the README.
