### 2017-11-02 f3bdcec
Add script to download and prepare MNIST #1
