### 2017-11-02 haythamfayek/MatDL#1 @haythamfayek
Agreed. Added script to download and prepare MNIST.
