### 2017-11-02 #413 @haythamfayek
Thanks @nirum for your time and feedback. I addressed both issues you raised: haythamfayek/MatDL#1 and haythamfayek/MatDL#2.
