### 2017-11-07 #413 @haythamfayek
Hi @cMadan

Here is the DOI from Zenodo:
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1042860.svg)](https://doi.org/10.5281/zenodo.1042860)

There is now a version mismatch from the above. The submitted version is 0.7 vs the archived version is 0.8. How can I amend that?
