### 2017-11-07 #413 @nirum
I updated this issue with the new version. In case you reference the version in the paper.md file, edit it there too--I think that's all you have to do.
