### 2017-11-07 #413 @arfon
> @arfon - the last line in the pdf has a link that overflows into the margin and out of the pdf. Is there anything that can be done about it?

Huh, yeah, something weird going on in the [LaTeX template](https://github.com/openjournals/whedon/blob/master/resources/latex.template) we use to produce the PDF. A couple of thoughts:

- The word 'GitHub' in that sentence already links to the GitHub repository in the PDF so perhaps we don't need the URL called out separately too.
- JOSS papers by default link to the software repository already (see the left-hand column) so this entire sentence could be considered redundant?
