### 2017-11-08 #413 @haythamfayek
@arfon - I am happy to remove the link from `paper.md` (Will that auto recompile?) but the links will continue to overflow in other documents as well. We can add the `breaklinks` option to the hyperref package to allow multi-line links. Would you like me to submit a pull request?
