### 2017-11-08 #413 @arfon
> Would you like me to submit a pull request?

Yes please do. I'll recompile the paper once you've made the change to the LaTeX template.
